requirements.txt
uvicorn
fastapi
motor
pydantic
dotenv
passlib
PyJWT
python-jose
httpx